requirements.txt

python-dotenv

# openpyxl==3.1.5
# XlsxWriter==3.2.2
# scipy  == 1.13.0
# fastapi == 0.115.11
# uvicorn == 0.30.6
# tiktoken == 0.5.1
# gradio == 5.23.2
# boto3>=1.37.23
# requests==2.32.3

APScheduler==3.11.0
python-telegram-bot==22.0
numpy==2.0.2
pandas==2.2.2
google-api-python-client==2.169.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
oauth2client==3.0.0
openai==1.102.0
# scikit-learn==1.6.1
nltk==3.9.1
tzlocal

boto3>=1.37.23
watchtower==3.4.0

python-docx
pdfminer.six

phonenumbers
langdetect
langcodes